% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickvis_multigeom.R
\name{quickvis_mulitgeom}
\alias{quickvis_mulitgeom}
\title{Allows to use multiple different geoms with different aesthetic mapping. Intended to just give an first impression how the different geoms would look as a plot. Lacks functionality compared to using the normal ggplot function. For a bit more functionality use \code{\link[=quickvis_simple]{quickvis_simple()}} and add further geoms after the function.}
\usage{
quickvis_mulitgeom(
  data = NULL,
  geom1 = c("x", "y", "geom"),
  geom2 = c("x", "y", "geom"),
  geom3 = c("x", "y", "geom")
)
}
\arguments{
\item{data}{Default dataset to use for plot.}

\item{geom1}{First of 3 geoms available to add. Plot as follows "c ("Mapping for x-axis", "Mapping for y- Axis", "geom").}

\item{geom2}{Plot as follows "c ("Mapping for x-axis", "Mapping for y- Axis", "geom").}

\item{geom3}{Plot as follows "c ("Mapping for x-axis", "Mapping for y- Axis", "geom").}
}
\value{
ggplot
}
\description{
Allows to use multiple different geoms with different aesthetic mapping. Intended to just give an first impression how the different geoms would look as a plot. Lacks functionality compared to using the normal ggplot function. For a bit more functionality use \code{\link[=quickvis_simple]{quickvis_simple()}} and add further geoms after the function.
}
\examples{
quickvis_mulitgeom(mtcars, c ("hp", "cyl", "point"), c ("hp", "cyl", "line"))
}
