% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickvis_simple.R
\name{quickvis_simple}
\alias{quickvis_simple}
\title{Alternative to just using ggplot if you want to create a simple plot to visualise data quicker with less complex code. Uses ggplot internally to create the plot.}
\usage{
quickvis_simple(
  data = NULL,
  x,
  y,
  geom = "point",
  theme = "none",
  facet_wrap = "F",
  col = NULL
)
}
\arguments{
\item{data}{Default dataset to use for plot.}

\item{x}{Variable which will be mapped to the x-axis of the plot.}

\item{y}{Variable which will be mapped to the y-axis of the plot.}

\item{geom}{Decides geom used. No input will result in the usage of \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{theme}{Lets you decide which theme you want to use. No theme will be applied if left empty.}

\item{facet_wrap}{Decides which variable you want to use for a facet_wrap. When left empty there will be no facet wrap.}

\item{col}{Decides variable of dataset which will color the geom. Not colors will be applied if left empty.}
}
\description{
Alternative to just using ggplot if you want to create a simple plot to visualise data quicker with less complex code. Uses ggplot internally to create the plot.
}
\examples{

quickvis_simple(mtcars, "hp", "cyl")
quickvis_simple(mtcars, "hp", geom = "bar")
quickvis_simple(mtcars, "mpg", "cyl", "line", "bw")

# using facet_wrap argument
quickvis_simple(mtcars, "cyl", "hp", facet_wrap = "gear")

# using col argument
quickvis_simple(mtcars, "gear", "hp", col = "vs")

# U can also add other ggplot2 functions after the function to further enhance the plot
quickvis_simple(mtcars, "wt", "cyl", "point", "minimal") +
geom_line() +
labs(title = "Title", x = "X", y = "Y")

}
