% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickvis_simple.R
\name{quickvis_simple}
\alias{quickvis_simple}
\title{Create point plot quickly.}
\usage{
quickvis_simple(
  data = NULL,
  x,
  y,
  geom = "point",
  theme = "none",
  facet_wrap = "F",
  color = NULL
)
}
\arguments{
\item{data}{dataset}

\item{x}{x-axis parameter}

\item{y}{y-axis parameter}

\item{theme}{decide theme}

\item{facet_wrap}{decide if facetwrap should be enabled}
}
\value{
ggplot
}
\description{
Create point plot quickly.
}
\examples{
quickvis_simple(mtcars, "mpg", "cyl")
}
